postcard/requirements.txt
Django==1.8.3
djangorestframework==3.2.2
MySQL-python==1.2.5
wheel==0.24.0
